+++ cuchem/cuchem/scripts/prepare_ChEMBL_approved_drugs_data.py
